requirements.txt
tensorflow==2.15.0
numpy==1.24.3
pandas==2.1.1
scikit-learn==1.3.2
matplotlib==3.8.0
plotly==5.18.0
streamlit==1.28.0
skyfield==1.46.0
pyswisseph==2.10.3
MetaTrader5==5.0.45